# [cxx-interop] Add initial benchmark to compare vector<uint32_t> sum i…