--- a/benchmark/cmake/modules/AddSwiftBenchmarkSuite.cmake
+++ b/benchmark/cmake/modules/AddSwiftBenchmarkSuite.cmake
@@ -484,6 +484,8 @@ function (swift_benchmark_compile_archopts)
       set(cxx_options "")
       if ("${module_name_path}" MATCHES ".*cxx-source/.*")
         list(APPEND cxx_options "-Xfrontend" "-enable-experimental-cxx-interop" "-I" "${srcdir}/utils/CxxTests/")
+        # FIXME: https://github.com/apple/swift/issues/61453
+        list(APPEND cxx_options "-Xfrontend" "-validate-tbd-against-ir=none")
       endif()
 
       if ("${bench_flags}" MATCHES "-whole-module.*")