--- a/benchmark/cxx-source/CxxVectorSum.swift
+++ b/benchmark/cxx-source/CxxVectorSum.swift
@@ -17,6 +17,10 @@ import TestsUtils
 import CxxStdlibPerformance
 import Cxx
 
+// FIXME: Linux needs fix for https://github.com/apple/swift/issues/61547.
+#if os(Linux)
+public let benchmarks: [BenchmarkInfo] = []
+#else
 public let benchmarks = [
   BenchmarkInfo(
       name: "CxxVecU32.sum.Cxx.rangedForLoop",
@@ -122,3 +126,4 @@ public func run_CxxVectorOfU32_Sum_Swift_Reduce(_ n: Int) {
 extension VectorOfU32.const_iterator : Equatable, UnsafeCxxInputIterator { }
 
 extension VectorOfU32: CxxSequence {}
+#endif