--- a/benchmark/utils/CxxTests/CxxStdlibPerformance.h
+++ b/benchmark/utils/CxxTests/CxxStdlibPerformance.h
@@ -25,8 +25,5 @@ inline uint32_t testVector32Sum(size_t vectorSize, size_t iters) {
     return sum;
 }
 
-template<class T>
-inline T next(const T& i) { return i + 1; }
-
-template<class T>
-inline bool cmp(const T &lhs, const T &rhs) { return lhs == rhs; }
+// FIXME: remove when the templated operator == is correctly bridged.
+inline bool operator ==(const VectorOfU32::const_iterator &lhs, const VectorOfU32::const_iterator &rhs) { return lhs.base() == rhs.base(); }