--- a/benchmark/utils/CxxTests/CxxStdlibPerformance.h
+++ b/benchmark/utils/CxxTests/CxxStdlibPerformance.h
@@ -5,13 +5,21 @@
 
 using VectorOfU32 = std::vector<uint32_t>;
 
-inline VectorOfU32 makeVector32(size_t size) {
-    auto result = VectorOfU32();
-    result.reserve(size);
-    for (size_t i = 0; i < size; ++i) {
-        result.push_back(uint32_t(i));
+static inline VectorOfU32 vec;
+
+inline void initVector(size_t size) {
+    if (!vec.empty()) {
+        return;
     }
-    return result;
+    vec.reserve(size);
+    for (size_t i = 0; i < size; ++i) {
+        vec.push_back(uint32_t(i));
+    }
+}
+
+inline VectorOfU32 makeVector32(size_t size) {
+    initVector(size);
+    return vec;
 }
 
 inline uint32_t testVector32Sum(size_t vectorSize, size_t iters) {