--- a/benchmark/utils/CxxTests/module.modulemap
+++ b/benchmark/utils/CxxTests/module.modulemap
@@ -7,3 +7,8 @@ module CxxSubscripts {
   header "Subscripts.h"
   requires cplusplus
 }
+
+module CxxStdlibPerformance {
+  header "CxxStdlibPerformance.h"
+  requires cplusplus
+}