…n C++ vs Swift

This benchmark compares the performance of summing up a vector of a million elements between Swift and C++.

Initial numbers on M1 Mac:

```
#,TEST,SAMPLES,MIN(μs),MAX(μs),MEAN(μs),SD(μs),MEDIAN(μs)
139,CreateObjects,200,11,29,14,2,15
140,CxxVectorOfU32SumInCxx,200,77,107,95,7,93
141,CxxVectorOfU32SumInSwift,36,24980,27957,25654,646,25546
142,CxxVectorOfU32SumInSwift_Fastest,200,77,83,77,1,77
```

We're investigating why the `CxxVectorOfU32SumInSwift` is slower and how to correctly fix it.