@hyp 2022-10-05
@swift-ci please test macOS platform