@zoecarver 2022-10-05
<details>
<summary>Benchmark Info</summary>
<br>

### Performance (x86_64): -O

**Regression**                    | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                              | ---:  | ---:  | ---:   | ---:     
Dictionary4                       | 153   | 193   | +26.1% | **0.79x (?)**
&nbsp; | | | | 
**Improvement**                   | **OLD** | **NEW** | **DELTA** | **RATIO**
FlattenListLoop                   | 1620  | 1386  | -14.4% | **1.17x (?)**
ObjectiveCBridgeStubToNSStringRef | 93    | 86    | -7.5%  | **1.08x (?)**
&nbsp; | | | | 
**Added**                         | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
CxxVectorOfU32SumInCxx            | 3     | 5     | 4      | —      
CxxVectorOfU32SumInSwift          | 38989 | 39456 | 39183  | —      
CxxVectorOfU32SumInSwift_Fastest  | 3     | 3     | 3      | —      

### Code size: -O


### Performance (x86_64): -Osize

**Regression**                   | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                             | ---:  | ---:  | ---:   | ---:     
PrefixAnySeqCntRange             | 13    | 16    | +23.1% | **0.81x (?)**
ObjectiveCBridgeStubDateAccess   | 130   | 152   | +16.9% | **0.86x (?)**
FlattenListLoop                  | 1385  | 1615  | +16.6% | **0.86x (?)**
PrefixWhileAnySequence           | 172   | 198   | +15.1% | **0.87x (?)**
StringComparison_ascii           | 360   | 390   | +8.3%  | **0.92x (?)**
&nbsp; | | | | 
**Improvement**                  | **OLD** | **NEW** | **DELTA** | **RATIO**
Dictionary4                      | 231   | 159   | -31.2% | **1.45x (?)**
Dictionary4OfObjects             | 316   | 277   | -12.3% | **1.14x (?)**
&nbsp; | | | | 
**Added**                        | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
CxxVectorOfU32SumInCxx           | 54    | 56    | 55     | —      
CxxVectorOfU32SumInSwift         | 37043 | 38907 | 38182  | —      
CxxVectorOfU32SumInSwift_Fastest | 35    | 36    | 35     | —      

### Code size: -Osize


### Performance (x86_64): -Onone

**Improvement**                         | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                                    | ---:  | ---:  | ---:  | ---:     
ConvertFloatingPoint.MockFloat64Exactly | 475   | 441   | -7.2% | **1.08x (?)**
&nbsp; | | | | 
**Added**                               | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
CxxVectorOfU32SumInCxx                  | 8147  | 8154  | 8152  | —      
CxxVectorOfU32SumInSwift                | 72724 | 77494 | 75066 | —      
CxxVectorOfU32SumInSwift_Fastest        | 10992 | 11758 | 11369 | —      

### Code size: -swiftlibs



✅  | Benchmark Check Report
---|---
⚠️🔤 | `CxxVectorOfU32SumInSwift` name is composed of 6 words. <br><sub> Split `CxxVectorOfU32SumInSwift` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️⏱ | `CxxVectorOfU32SumInSwift` has setup overhead of 36214 μs (99.6%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⚠️Ⓜ️ | `CxxVectorOfU32SumInSwift` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32SumInSwift` mem_pages [i1, i2]: min=[3928, 3928] 𝚫=0 R=[5876, 2938]
⚠️🔤 | `CxxVectorOfU32SumInCxx` name is composed of 6 words. <br><sub> Split `CxxVectorOfU32SumInCxx` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️⏱ | `CxxVectorOfU32SumInCxx` has setup overhead of 70 μs (100.0%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⛔️⏱ | `CxxVectorOfU32SumInCxx` execution took 0 μs. <br><sub> Ensure the workload of `CxxVectorOfU32SumInCxx` has a properly measurable size (runtime > 20 μs) and is not eliminated by the compiler (use `blackHole` function if necessary).
⚠️Ⓜ️ | `CxxVectorOfU32SumInCxx` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32SumInCxx` mem_pages [i1, i2]: min=[1960, 1960] 𝚫=0 R=[0, 19]
⛔️🔤 | `CxxVectorOfU32SumInSwift_Fastest` name doesn`t conform to benchmark naming convention. <br><sub> See http://bit.ly/BenchmarkNaming
⚠️🔤 | `CxxVectorOfU32SumInSwift_Fastest` name is composed of 6 words. <br><sub> Split `CxxVectorOfU32SumInSwift_Fastest` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️⏱ | `CxxVectorOfU32SumInSwift_Fastest` has setup overhead of 78 μs (106.8%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⚠️⏱ | `CxxVectorOfU32SumInSwift_Fastest` execution took -5 μs. <br><sub> Increase the workload of `CxxVectorOfU32SumInSwift_Fastest` to be more than 20 μs.


<details>
  <summary><strong>How to read the data</strong></summary>
The tables contain differences in performance which are larger than 8% and
differences in code size which are larger than 1%.

If you see any unexpected regressions, you should consider fixing the
regressions before you merge the PR.

**Noise**: Sometimes the performance results (not code size!) contain false
alarms. Unexpected regressions which are marked with '(?)' are probably noise.
If you see regressions which you cannot explain you can try to run the
benchmarks again. If regressions still show up, please consult with the
performance team (@eeckstein).
</details>

<details>
  <summary><strong>Hardware Overview</strong></summary>

      Model Name: Mac mini
      Model Identifier: Macmini8,1
      Processor Name: 6-Core Intel Core i7
      Processor Speed: 3.2 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache (per Core): 256 KB
      L3 Cache: 12 MB
      Memory: 32 GB

</details>

</details>