@hyp 2022-10-07
@swift-ci please benchmark