@hyp 2022-10-07
Benchmark info:


### Performance (x86_64): -O

**Regression**                                 | **OLD**  | **NEW**  | **DELTA** | **RATIO**
:---                                           | ---:     | ---:     | ---:     | ---:     
StringInterpolationManySmallSegments           | 7300     | 8800     | +20.5%   | **0.83x (?)**
Dictionary3                                    | 122      | 146      | +19.7%   | **0.84x (?)**
RemoveWhereMoveInts                            | 12       | 13       | +8.3%    | **0.92x (?)**
&nbsp; | | | | 
**Improvement**                                | **OLD**  | **NEW**  | **DELTA** | **RATIO**
Dictionary4                                    | 201      | 152      | -24.4%   | **1.32x**
Dictionary4OfObjects                           | 248      | 209      | -15.7%   | **1.19x (?)**
&nbsp; | | | | 
**Added**                                      | **MIN**  | **MAX**  | **MEAN** | **MAX_RSS**
CxxVectorOfU32.Sum.Cxx.RangedForLoop           | 101      | 103      | 102      | —      
CxxVectorOfU32.Sum.Swift.ForInLoop             | 36881    | 41604    | 39568    | —      
CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop | 10690028 | 11153274 | 10972647 | —      
CxxVectorOfU32.Sum.Swift.RawIteratorLoop       | 90       | 91       | 91       | —      
CxxVectorOfU32.Sum.Swift.Reduce                | 39856    | 40399    | 40106    | —      

### Code size: -O


### Performance (x86_64): -Osize

**Regression**                                 | **OLD**  | **NEW**  | **DELTA** | **RATIO**
:---                                           | ---:     | ---:     | ---:     | ---:     
PrefixAnySeqCRangeIter                         | 13       | 15       | +15.4%   | **0.87x (?)**
StringWalk                                     | 1360     | 1520     | +11.8%   | **0.89x (?)**
PrefixAnySeqCntRange                           | 13       | 14       | +7.7%    | **0.93x (?)**
&nbsp; | | | | 
**Improvement**                                | **OLD**  | **NEW**  | **DELTA** | **RATIO**
ObjectiveCBridgeStubDateAccess                 | 152      | 130      | -14.5%   | **1.17x (?)**
DictionaryKeysContainsCocoa                    | 15       | 14       | -6.7%    | **1.07x (?)**
&nbsp; | | | | 
**Added**                                      | **MIN**  | **MAX**  | **MEAN** | **MAX_RSS**
CxxVectorOfU32.Sum.Cxx.RangedForLoop           | 331      | 332      | 331      | —      
CxxVectorOfU32.Sum.Swift.ForInLoop             | 36610    | 38637    | 37663    | —      
CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop | 10746758 | 10871907 | 10796945 | —      
CxxVectorOfU32.Sum.Swift.RawIteratorLoop       | 325      | 329      | 327      | —      
CxxVectorOfU32.Sum.Swift.Reduce                | 39519    | 39997    | 39750    | —      

### Code size: -Osize


### Performance (x86_64): -Onone

**Regression**                                 | **OLD**   | **NEW**   | **DELTA** | **RATIO**
:---                                           | ---:      | ---:      | ---:      | ---:     
ObjectiveCBridgeStubNSDateRefAccess            | 2516      | 3057      | +21.5%    | **0.82x (?)**
ObjectiveCBridgeStubDateAccess                 | 2604      | 3077      | +18.2%    | **0.85x (?)**
NSError                                        | 371       | 403       | +8.6%     | **0.92x (?)**
&nbsp; | | | | 
**Improvement**                                | **OLD**   | **NEW**   | **DELTA** | **RATIO**
ObjectiveCBridgeStubDateMutation               | 352       | 326       | -7.4%     | **1.08x (?)**
&nbsp; | | | | 
**Added**                                      | **MIN**   | **MAX**   | **MEAN**  | **MAX_RSS**
CxxVectorOfU32.Sum.Cxx.RangedForLoop           | 11179     | 11419     | 11336     | —      
CxxVectorOfU32.Sum.Swift.ForInLoop             | 67463     | 67624     | 67524     | —      
CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop | 703459153 | 704797431 | 704231260 | —      
CxxVectorOfU32.Sum.Swift.RawIteratorLoop       | 14508     | 14595     | 14548     | —      
CxxVectorOfU32.Sum.Swift.Reduce                | 80136     | 81468     | 80918     | —      

### Code size: -swiftlibs



✅  | Benchmark Check Report
---|---
⚠️🔤 | `CxxVectorOfU32.Sum.Cxx.RangedForLoop` name is composed of 6 words. <br><sub> Split `CxxVectorOfU32.Sum.Cxx.RangedForLoop` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️⏱ | `CxxVectorOfU32.Sum.Cxx.RangedForLoop` has setup overhead of 122 μs (64.2%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⚠️Ⓜ️ | `CxxVectorOfU32.Sum.Cxx.RangedForLoop` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32.Sum.Cxx.RangedForLoop` mem_pages [i1, i2]: min=[102, 102] 𝚫=0 R=[105, 210]
⚠️🔤 | `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` name is composed of 7 words. <br><sub> Split `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️🔤 | `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` name is 46 characters long. <br><sub> Benchmark name should not be longer than 40 characters.
⛔️⏱ | `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` execution took at least 10495803 μs. <br><sub> Decrease the workload of `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` by a factor of 16384 (100000), to be less than 1000 μs.
⚠️Ⓜ️ | `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32.Sum.Swift.IndexAndSubscriptLoop` mem_pages [i1, i2]: min=[1018, 1020] 𝚫=2 R=[206, 102]
⚠️🔤 | `CxxVectorOfU32.Sum.Swift.ForInLoop` name is composed of 6 words. <br><sub> Split `CxxVectorOfU32.Sum.Swift.ForInLoop` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️⏱ | `CxxVectorOfU32.Sum.Swift.ForInLoop` execution took at least 34850 μs. <br><sub> Decrease the workload of `CxxVectorOfU32.Sum.Swift.ForInLoop` by a factor of 64 (100), to be less than 1000 μs.
⚠️Ⓜ️ | `CxxVectorOfU32.Sum.Swift.ForInLoop` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32.Sum.Swift.ForInLoop` mem_pages [i1, i2]: min=[299, 299] 𝚫=0 R=[1211, 301]
⛔️⏱ | `CxxVectorOfU32.Sum.Swift.Reduce` execution took at least 36683 μs. <br><sub> Decrease the workload of `CxxVectorOfU32.Sum.Swift.Reduce` by a factor of 64 (100), to be less than 1000 μs.
⚠️Ⓜ️ | `CxxVectorOfU32.Sum.Swift.Reduce` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32.Sum.Swift.Reduce` mem_pages [i1, i2]: min=[300, 300] 𝚫=0 R=[917, 1204]
⚠️🔤 | `CxxVectorOfU32.Sum.Swift.RawIteratorLoop` name is composed of 6 words. <br><sub> Split `CxxVectorOfU32.Sum.Swift.RawIteratorLoop` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⛔️⏱ | `CxxVectorOfU32.Sum.Swift.RawIteratorLoop` has setup overhead of 126 μs (66.0%). <br><sub> Move initialization of benchmark data to the `setUpFunction` registered in `BenchmarkInfo`.
⚠️Ⓜ️ | `CxxVectorOfU32.Sum.Swift.RawIteratorLoop` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVectorOfU32.Sum.Swift.RawIteratorLoop` mem_pages [i1, i2]: min=[102, 102] 𝚫=0 R=[105, 105]