@hyp 2022-10-10
<details>
<summary>benchmark info</summary>
### Performance (x86_64): -O

**Regression**                    | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                              | ---: | ---: | ---:   | ---:     
Dictionary3                       | 124  | 154  | +24.2% | **0.81x (?)**
DropFirstArray                    | 13   | 14   | +7.7%  | **0.93x (?)**
&nbsp; | | | | 
**Improvement**                   | **OLD** | **NEW** | **DELTA** | **RATIO**
Dictionary4                       | 227  | 152  | -33.0% | **1.49x**
Dictionary4OfObjects              | 255  | 210  | -17.6% | **1.21x**
Fibonacci                         | 6    | 5    | -16.7% | **1.20x (?)**
TypeName                          | 673  | 620  | -7.9%  | **1.09x (?)**
RemoveWhereFilterInts             | 27   | 25   | -7.4%  | **1.08x**
DictionaryKeysContainsNative      | 15   | 14   | -6.7%  | **1.07x (?)**
DropLastAnySeqCRangeIter          | 381  | 356  | -6.6%  | **1.07x (?)**
&nbsp; | | | | 
**Added**                         | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
CxxVecU32.Sum.Cxx.RangedForLoop   | 10   | 10   | 10     | —      
CxxVecU32.Sum.Swift.ForInLoop     | 6185 | 6398 | 6316   | —      
CxxVecU32.Sum.Swift.IteratorLoop  | 10   | 10   | 10     | —      
CxxVecU32.Sum.Swift.Reduce        | 6376 | 6502 | 6432   | —      
CxxVecU32.Sum.Swift.SubscriptLoop | 10   | 10   | 10     | —      

### Code size: -O


### Performance (x86_64): -Osize

**Regression**                    | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                              | ---: | ---: | ---:   | ---:     
StringWalk                        | 1240 | 1520 | +22.6% | **0.82x (?)**
StrComplexWalk                    | 2800 | 3100 | +10.7% | **0.90x (?)**
&nbsp; | | | | 
**Improvement**                   | **OLD** | **NEW** | **DELTA** | **RATIO**
ObjectiveCBridgeStubDateAccess    | 152  | 130  | -14.5% | **1.17x (?)**
&nbsp; | | | | 
**Added**                         | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
CxxVecU32.Sum.Cxx.RangedForLoop   | 78   | 78   | 78     | —      
CxxVecU32.Sum.Swift.ForInLoop     | 6382 | 6490 | 6433   | —      
CxxVecU32.Sum.Swift.IteratorLoop  | 78   | 80   | 79     | —      
CxxVecU32.Sum.Swift.Reduce        | 6622 | 6719 | 6656   | —      
CxxVecU32.Sum.Swift.SubscriptLoop | 78   | 78   | 78     | —      

### Code size: -Osize


### Performance (x86_64): -Onone

**Regression**                           | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                                     | ---:  | ---:  | ---:  | ---:     
SubstringFromLongString2                 | 88    | 95    | +8.0% | **0.93x (?)**
&nbsp; | | | | 
**Improvement**                          | **OLD** | **NEW** | **DELTA** | **RATIO**
ObjectiveCBridgeStubFromArrayOfNSString2 | 1670  | 1550  | -7.2% | **1.08x (?)**
&nbsp; | | | | 
**Added**                                | **MIN** | **MAX** | **MEAN** | **MAX_RSS**
CxxVecU32.Sum.Cxx.RangedForLoop          | 1587  | 1623  | 1600  | —      
CxxVecU32.Sum.Swift.ForInLoop            | 10684 | 10814 | 10750 | —      
CxxVecU32.Sum.Swift.IteratorLoop         | 1784  | 1827  | 1799  | —      
CxxVecU32.Sum.Swift.Reduce               | 12520 | 12689 | 12617 | —      
CxxVecU32.Sum.Swift.SubscriptLoop        | 1784  | 1816  | 1795  | —      

### Code size: -swiftlibs



✅  | Benchmark Check Report
---|---
⚠️⏱ | `CxxVecU32.Sum.Swift.SubscriptLoop` execution took 11 μs. <br><sub> Increase the workload of `CxxVecU32.Sum.Swift.SubscriptLoop` to be more than 20 μs.
⚠️🔤 | `CxxVecU32.Sum.Swift.ForInLoop` name is composed of 5 words. <br><sub> Split `CxxVecU32.Sum.Swift.ForInLoop` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⚠️⏱ | `CxxVecU32.Sum.Swift.ForInLoop` execution took at least 6038 μs. <br><sub> Decrease the workload of `CxxVecU32.Sum.Swift.ForInLoop` by a factor of 8 (10), to be less than 1000 μs.
⚠️Ⓜ️ | `CxxVecU32.Sum.Swift.ForInLoop` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVecU32.Sum.Swift.ForInLoop` mem_pages [i1, i2]: min=[161, 79] 𝚫=82 R=[103, 324]
⚠️⏱ | `CxxVecU32.Sum.Swift.IteratorLoop` execution took 10 μs. <br><sub> Increase the workload of `CxxVecU32.Sum.Swift.IteratorLoop` to be more than 20 μs.
⚠️Ⓜ️ | `CxxVecU32.Sum.Swift.IteratorLoop` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVecU32.Sum.Swift.IteratorLoop` mem_pages [i1, i2]: min=[28, 28] 𝚫=0 R=[0, 32]
⚠️🔤 | `CxxVecU32.Sum.Cxx.RangedForLoop` name is composed of 5 words. <br><sub> Split `CxxVecU32.Sum.Cxx.RangedForLoop` name into dot-separated groups and variants. See http://bit.ly/BenchmarkNaming
⚠️⏱ | `CxxVecU32.Sum.Cxx.RangedForLoop` execution took 11 μs. <br><sub> Increase the workload of `CxxVecU32.Sum.Cxx.RangedForLoop` to be more than 20 μs.
⚠️⏱ | `CxxVecU32.Sum.Swift.Reduce` execution took at least 6315 μs. <br><sub> Decrease the workload of `CxxVecU32.Sum.Swift.Reduce` by a factor of 8 (10), to be less than 1000 μs.
⚠️Ⓜ️ | `CxxVecU32.Sum.Swift.Reduce` has very wide range of memory used between independent, repeated measurements. <br><sub> `CxxVecU32.Sum.Swift.Reduce` mem_pages [i1, i2]: min=[80, 80] 𝚫=0 R=[164, 328]




</details>